+++ docs/docs/99-api/examples/automation_demo/more_specific_demos/demo_B.js
@@ -11,18 +11,16 @@ const { MemoryModel, drawAutomatedStackFrames } = require("memory-viz");
 const WIDTH = 1300;
 const listOfStackFrames = [
     {
-        isClass: true,
+        type: ".frame",
         name: "__main__",
         id: null,
         value: { lst1: 82, lst2: 84, p: 99, d: 10, t: 11 },
-        stack_frame: true,
     },
     {
-        isClass: true,
+        type: ".frame",
         name: "Animal",
         id: null,
         value: { age: 2, name: 94 },
-        stack_frame: true,
     },
 ];