+++ memory-viz/src/style.ts
@@ -88,10 +88,10 @@ function populateStyleObject(object: DrawnEntity, seed: Number) {
         object_type = "primitive";
     } else if (collections.includes(object.type)) {
         object_type = "collection";
-    } else if (object.stack_frame) {
-        object_type = "stackframe";
-    } else {
+    } else if (object.type === ".class") {
         object_type = "class";
+    } else {
+        object_type = "stackframe";
     }
 
     // We then add properties specific to the different type categories.